Move main.go to cmd/
