+++ cmd/main.go
@@ -16,10 +16,8 @@ limitations under the License.
 
 package main
 
-import (
-	"github.com/3scale-ops/marin3r/cmd"
-)
+import "github.com/3scale-ops/marin3r/cmd/app"
 
 func main() {
-	cmd.Execute()
+	app.Execute()
 }
